Initial commit.
This serves to scratch my initial itch of "Can I automate an
installation of a load-balanced Puppet Enterprise environment using
Vagrant", but isn't really all that useful for testing Puppet Code in
that Code Manager or stand-alone File Sync haven't been configured, so
no code changes will reach the Compile Masters without manual
intervention.

This is possibly something to look at later, but the main purpose of all
this was to give me a test environment for modifying or updating Puppet
infrastructure.