Clarify comment about using vbguest auto_update.
'centos/7' isn't subject to the same issues that affect RHEL, so maybe
you want to automatically install the VirtualBox Guest Additions as they
don't come with the 'centos/7' box.

Also, 'centos/7' syncs the '.' directory to '/vagrant' using rsync
instead of mounting the directory due to the before-mentioned lack of
Guest Additions.  So if you are automatically installing/updating the
guest additions you might want to uncomment the line that enables
mounting of '.' into '/vagrant'.